Move all network related hooks together
