+++ src/components/pages/create-pool-page/components/CreatePoolSuccessModal/CreatePoolSuccessModal.tsx
@@ -4,7 +4,7 @@ import ActionButton from "@/src/components/common/ActionButton/ActionButton";
 import {useCallback} from "react";
 import {openNewTab} from "@/src/utils/common";
 import {CoinName} from "@/src/utils/coinsConfig";
-import useAppUrl from "@/src/hooks/useAppUrl";
+import useAppUrl from "@/src/hooks/network/useAppUrl";
 
 type Props = {
   coinA: CoinName;
