+++ src/hooks/network/useAppUrl.ts
@@ -1,7 +1,7 @@
 import {CHAIN_IDS} from "fuels";
 import {useProvider} from "@fuels/react";
 
-import {FUEL_APP_URL_MAP} from "../utils/constants";
+import {FUEL_APP_URL_MAP} from "../../utils/constants";
 
 const defaultValue = FUEL_APP_URL_MAP.get(CHAIN_IDS.fuel.mainnet) ?? "";
