+++ src/hooks/network/useExplorerBaseUrl.ts
@@ -1,7 +1,7 @@
 import {CHAIN_IDS} from "fuels";
 import {useProvider} from "@fuels/react";
 
-import {EXPLORER_BASE_URL_MAP} from "../utils/constants";
+import {EXPLORER_BASE_URL_MAP} from "../../utils/constants";
 
 const defaultValue = EXPLORER_BASE_URL_MAP.get(CHAIN_IDS.fuel.mainnet) ?? "";
