+++ src/hooks/network/useSQDIndexerUrl.ts
@@ -1,7 +1,7 @@
 import {CHAIN_IDS} from "fuels";
 import {useProvider} from "@fuels/react";
 
-import {SQD_INDEXER_URL_MAP} from "../utils/constants";
+import {SQD_INDEXER_URL_MAP} from "../../utils/constants";
 
 const defaultValue = SQD_INDEXER_URL_MAP.get(CHAIN_IDS.fuel.mainnet) ?? "";
